# swagger.json
{"openapi":"3.0.0","info":{"title":"Node S15","version":"1.0.0","description":"This is a simple CRUD API","license":{"name":"MIT","url":"http://mit.com"},"contact":{"name":"Juan Carlos Fernandez","url":"https://github.com/Juancfer","email":"juan@example.com"}},"server":[{"url":"http://localhost:3000"}],"paths":{"/author":{"get":{"summary":"Lists all the authors","tags":["Author"],"responses":{"200":{"description":"The list of the authors","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Author"}},"pagination":{"$ref":"#/components/schemas/Pagination"}}}}}}}},"post":{"summary":"Create a new author","tags":["Author"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"responses":{"201":{"description":"The created author","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"404":{"description":"The request body is incorrect or missing"}}}},"/author/{id}":{"get":{"summary":"Get author by ID","tags":["Author"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"ID of the author"}],"responses":{"200":{"description":"The author","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"404":{"description":"Author not found"}}},"delete":{"summary":"Delete author by ID","tags":["Author"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"ID of the author"}],"responses":{"200":{"description":"The deleted author","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"401":{"description":"Unauthorized access"},"404":{"description":"Author not found"}}},"put":{"summary":"Update author by ID","tags":["Author"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"ID of the author"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"responses":{"200":{"description":"The updated author","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"401":{"description":"Unauthorized access"},"404":{"description":"Author not found"}}}},"/author/name/{name}":{"get":{"summary":"Get authors by name","tags":["Author"],"parameters":[{"in":"path","name":"name","schema":{"type":"string"},"required":true,"description":"Name of the author"}],"responses":{"200":{"description":"The authors","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Author"}}}}},"404":{"description":"Authors not found"}}}},"/author/image-upload":{"post":{"summary":"Upload author image","tags":["Author"],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"authorId":{"type":"string"},"logo":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"The updated author with the new image","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}}},"404":{"description":"Author not found"}}}},"/author/login":{"post":{"summary":"Login as an author","tags":["Author"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"Login successful, returns JWT token","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string"}}}}}},"400":{"description":"Missing email or password"},"401":{"description":"Invalid credentials"}}}},"/book":{"get":{"summary":"Lists all the books","tags":["Book"],"responses":{"200":{"description":"The list of the books","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/Book"}},"pagination":{"$ref":"#/components/schemas/Pagination"}}}}}}}},"post":{"summary":"Create a new book","tags":["Book"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"responses":{"201":{"description":"The created book","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"404":{"description":"The request body is incorrect or missing"}}}},"/book/{id}":{"get":{"summary":"Get book by ID","tags":["Book"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"ID of the book"}],"responses":{"200":{"description":"The book","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"404":{"description":"Book not found"}}},"delete":{"summary":"Delete book by ID","tags":["Book"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"ID of the book"}],"responses":{"200":{"description":"The deleted book","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"401":{"description":"Unauthorized access"},"404":{"description":"Book not found"}}},"put":{"summary":"Update book by ID","tags":["Book"],"parameters":[{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"ID of the book"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"responses":{"200":{"description":"The updated book","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}}},"401":{"description":"Unauthorized access"},"404":{"description":"Book not found"}}}},"/book/book/{book}":{"get":{"summary":"Get books by title","tags":["Book"],"parameters":[{"in":"path","name":"book","schema":{"type":"string"},"required":true,"description":"Title of the book"}],"responses":{"200":{"description":"The books","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}}},"404":{"description":"Books not found"}}}}},"components":{"schemas":{"Author":{"type":"object","required":["name"],"properties":{"name":{"type":"string","description":"Name of the Author (sample Ford)"},"email":{"type":"string","description":"Email of the author (sample author@mail.com)"},"password":{"type":"string","description":"Password of the author (sample 12345678)"},"country":{"type":"string","description":"Country of this Author (sample SPAIN)"},"profileImage":{"type":"string","description":"Profile image of this Author (sample http://balbblabla.com/image.jpg)"}}},"Book":{"type":"object","required":["title","pages"],"properties":{"title":{"type":"string","description":"Title of the book (sample Book Title)"},"author":{"type":"string","description":"ID of the author associated with the book"},"pages":{"type":"integer","description":"Number of pages in the book"},"publisher":{"type":"object","properties":{"name":{"type":"string","description":"Name of the publisher (sample Publisher Name)"},"country":{"type":"string","description":"Country of the publisher (sample SPAIN)"}}}}},"Pagination":{"type":"object","properties":{"totalItems":{"type":"number","description":"Total items on this endpoint"},"totalPages":{"type":"number","description":"Number of pages for this pagination"},"currentPage":{"type":"number","description":"Current page showed"}}}}},"tags":[{"name":"Author","description":"The Author managing API"},{"name":"Book","description":"The Book managing API"}]}